Refactored ingestion pages
