+++ app/apprentice-data-ingestion/page.tsx
@@ -1,17 +1,10 @@
-"use client"
-
-import type React from "react"
-
-import { useState } from "react"
+import React from "react";
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card"
 import { Button } from "@/components/ui/button"
-import { Badge } from "@/components/ui/badge"
-import { Progress } from "@/components/ui/progress"
-import { Upload, FileText, Database, CheckCircle, AlertCircle, Clock, RefreshCw, Download } from "lucide-react"
+import { CheckCircle, AlertCircle, Clock, Download } from "lucide-react";
+import ApprenticeUpsert from "@/components/apprentices/apprentice-upsert";
 
 export default function ApprenticeDataIngestionPage() {
-  const [uploadProgress, setUploadProgress] = useState(0)
-  const [isUploading, setIsUploading] = useState(false)
 
   const recentImports = [
     {
@@ -40,26 +33,6 @@ export default function ApprenticeDataIngestionPage() {
     },
   ]
 
-  const handleFileUpload = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const file = event.target.files?.[0]
-    if (file) {
-      setIsUploading(true)
-      setUploadProgress(0)
-
-      // Simulate upload progress
-      const interval = setInterval(() => {
-        setUploadProgress((prev) => {
-          if (prev >= 100) {
-            clearInterval(interval)
-            setIsUploading(false)
-            return 100
-          }
-          return prev + 10
-        })
-      }, 200)
-    }
-  }
-
   const getStatusIcon = (status: string) => {
     switch (status) {
       case "success":
@@ -74,20 +47,6 @@ export default function ApprenticeDataIngestionPage() {
     }
   }
 
-  const getStatusBadge = (status: string) => {
-    switch (status) {
-      case "success":
-      case "completed":
-        return <Badge className="bg-green-100 text-green-800">Success</Badge>
-      case "warning":
-        return <Badge className="bg-yellow-100 text-yellow-800">Warning</Badge>
-      case "failed":
-        return <Badge className="bg-red-100 text-red-800">Failed</Badge>
-      default:
-        return <Badge className="bg-gray-100 text-gray-800">Pending</Badge>
-    }
-  }
-
   return (
     <div className="p-6 space-y-6">
       <div>
@@ -95,47 +54,7 @@ export default function ApprenticeDataIngestionPage() {
         <p className="text-gray-600 mt-2">Import and manage apprentice data from various sources</p>
       </div>
 
-      {/* Upload Section */}
-      <Card>
-        <CardHeader>
-          <CardTitle className="flex items-center">
-            <Upload className="mr-2 h-5 w-5" />
-            Upload Apprentice CSV File
-          </CardTitle>
-          <CardDescription>Upload a CSV file containing apprentice data</CardDescription>
-        </CardHeader>
-        <CardContent>
-          <div className="border-2 border-dashed border-gray-300 rounded-lg p-6 text-center">
-            <Upload className="mx-auto h-12 w-12 text-gray-400" />
-            <div className="mt-4">
-              <label htmlFor="file-upload" className="cursor-pointer">
-                <span className="mt-2 block text-sm font-medium text-gray-900">
-                  Drop apprentice files here or click to upload
-                </span>
-                <input
-                  id="file-upload"
-                  name="file-upload"
-                  type="file"
-                  accept=".csv"
-                  className="sr-only"
-                  onChange={handleFileUpload}
-                />
-              </label>
-              <p className="mt-1 text-xs text-gray-500">CSV files up to 10MB</p>
-            </div>
-          </div>
-
-          {isUploading && (
-            <div className="mt-4">
-              <div className="flex justify-between text-sm text-gray-600 mb-2">
-                <span>Uploading...</span>
-                <span>{uploadProgress}%</span>
-              </div>
-              <Progress value={uploadProgress} className="w-full" />
-            </div>
-          )}
-        </CardContent>
-      </Card>
+      <ApprenticeUpsert />
 
       {/* Recent Imports */}
       <Card>
